Refactored: Grid and GridWrapper relocated
